bookshop/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-address==0.2.8
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
exceptiongroup==1.1.0
flake8==6.0.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pytest==7.2.1
python3-openid==3.2.0
pytz==2022.7
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.14
pip~=22.3.1
wheel~=0.37.1
Faker~=16.6.0
python-dateutil~=2.8.2
setuptools~=60.2.0
factory_boy==3.2.1
pytest-django==4.5.2
dj-database-url